refactor: split into components and containers
